card date to long and endpoints to dele through animalid

+++ ProductsApi/Models/Card.cs
@@ -5,7 +5,7 @@ namespace ProductsApi.Models
     public class Card
     {
         public int Id { get; set; }
-        public int Date { get; set; }
+        public long Date { get; set; }
         public string Status { get; set; } = string.Empty;
         public int AnimalId { get; set; }
